Modifications dotdev , optimisation code, commentaires

+++ backend/controllers/sauce.js
@@ -1,6 +1,9 @@
+// permet d'importer le modèle de données pour une sauce
 const Sauce = require("../models/Sauce");
+// permet d'importer le package file system
 const fs = require("fs");
 
+//permet de créer une nouvelle sauce
 exports.createSauce = (req, res, next) => {
 	const sauceObject = JSON.parse(req.body.sauce);
 	delete sauceObject._id;
@@ -15,12 +18,14 @@ exports.createSauce = (req, res, next) => {
 		.catch((error) => res.status(400).json({ error }));
 };
 
+//permet d'afficher une sauce
 exports.getOneSauce = (req, res, next) => {
 	Sauce.findOne({ _id: req.params.id })
 		.then((sauce) => res.status(200).json(sauce))
 		.catch((error) => res.status(404).json({ error }));
 };
 
+//permet de liker ou de disliker une sauce
 exports.likeSauce = (req, res, next) => {
 	const sauceObject = req.body;
 	const userId = sauceObject.userId;
@@ -67,6 +72,7 @@ exports.likeSauce = (req, res, next) => {
 		.catch((error) => res.status(400).json({ error }));
 };
 
+//permet de modifier une sauce
 exports.modifySauce = (req, res, next) => {
 	const sauceObject = req.file
 		? {
@@ -74,24 +80,39 @@ exports.modifySauce = (req, res, next) => {
 				imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`,
 		  }
 		: { ...req.body };
-	Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-		.then(() => res.status(200).json({ message: "Sauce modifiée !" }))
-		.catch((error) => res.status(400).json({ error }));
-};
-
-exports.deleteSauce = (req, res, next) => {
 	Sauce.findOne({ _id: req.params.id })
 		.then((sauce) => {
-			const filename = sauce.imageUrl.split("/images/")[1];
-			fs.unlink(`images/${filename}`, () => {
-				Sauce.deleteOne({ _id: req.params.id })
-					.then(() => res.status(200).json({ message: "Sauce supprimée !" }))
-					.catch((error) => res.status(400).json({ error }));
-			});
+			// vérification que sa soit bien l'utilisateur qui a creer la sauce
+			if (sauce.userId === req.userIdAuth) {
+				const filename = sauce.imageUrl.split("/images/")[1];
+				fs.unlink(`images/${filename}`, () => {
+					Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+						.then(() => res.status(200).json({ message: "Objet modifié !" }))
+						.catch((error) => res.status(400).json({ error }));
+				});
+			}
 		})
 		.catch((error) => res.status(500).json({ error }));
 };
 
+//permet de supprimer une sauce
+exports.deleteSauce = (req, res, next) => {
+	Sauce.findOne({ _id: req.params.id })
+		.then((sauce) => {
+			// vérification que sa soit bien l'utilisateur qui a creer la sauce
+			if (sauce.userId === req.userIdAuth) {
+				const filename = sauce.imageUrl.split("/images/")[1];
+				fs.unlink(`images/${filename}`, () => {
+					Sauce.deleteOne({ _id: req.params.id })
+						.then(() => res.status(200).json({ message: "Sauce supprimée !" }))
+						.catch((error) => res.status(400).json({ error }));
+				});
+			}
+		})
+		.catch((error) => res.status(500).json({ error }));
+};
+
+//permet d'afficher toutes les sauces
 exports.getAllSauce = (req, res, next) => {
 	//res.status(200).json({ message: 'Votre requête a bien été reçue !' });
 	Sauce.find()
